auto_correction_service/requirements.txt
absl-py==0.7.1
astor==0.8.0
bleach==1.5.0
captcha==0.2.4
#certifi==2016.2.28
certifi
cloudpickle==1.2.1
cycler==0.10.0
Cython==0.29.12
dask==2.1.0
decorator==4.4.0
easydict==1.8
elastic-apm==4.2.2
entrypoints==0.2.3
gast==0.2.2
grpcio==1.22.0
grpcio-tools==1.22.0
h5py==2.8.0
html5lib==0.9999999
kiwisolver==1.1.0
logger==1.4
Markdown==3.1.1
matplotlib==2.2.3
mpmath==1.1.0
networkx==2.3
numpy==1.15.1
opencv-python==3.4.2.17
Pillow==5.2.0
protobuf==3.6.1
psutil==5.6.3
pyparsing==2.4.0
python-dateutil==2.7.3
python-Levenshtein==0.12.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
Shapely==1.6.4.post2
six==1.12.0
sympy==1.4
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
toolz==0.9.0
tqdm==4.25.0
urllib3==1.25.3
Werkzeug==0.15.4
